Start work on general spline
This commit adds an API for general splines with arbitrary mixing of
"auto" and "given" points, using a spline builder interface.

At this point, it works for simple interpolating splines, but hasn't
been tested for the general case, and some functionality is missing,
notably the ability to solve for G2 continuity by setting the tension
(bias) of an auto point mirrored with a given point.