Make SplineSpec reusable
Currently, generating a Spline from a SplineSpec consumes
the SplineSpec, moving the segments in to the Spline.

With this change, the Spline now borrows the segments, although
it can easily take ownership via Spline::into_owned. The SplineSpec
can then be modified, either by adding existing elements or by
mutating elements in place.

The latter option should be done with care, as it could allow
the elements to be put in an invalid state.

The motivation here is to allow efficient updating of the spline,
such as in an interactive editing tool.