Improve rendering
This replaces the fixed subdivision with a dynamic approach, and also
changes the parametrization to do a better job when tension goes up.
It's a bit hacky and heuristic, rather than being based on careful
tuning of parameters to approach an optimized rendering, but it should
be a big improvement over the current situation.

Fixes #15

Also closes #12 while I'm deep in the curve math.